@@ -4,7 +4,7 @@
   "description": "",
   "main": "main.js",
   "engines": {
-    "node": "10.16.3"
+    "node": "16.13.1"
   },
   "scripts": {
     "lint": "eslint .",
